+++ src/scitacean/testing/strategies.py
@@ -106,12 +106,18 @@ def datasets(
             if field.name != "type"
         }
 
+    try:
+        pid = make_fixed_arg("pid")
+        del fixed["pid"]
+    except KeyError:
+        pid = st.builds(PID)
+
     kwargs = make_args(read_only=False)
     read_only_arg = st.fixed_dictionaries(make_args(read_only=True))
     return st.builds(
         Dataset,
         type=st.just(dataset_type),
         _read_only=read_only_arg,
-        _pid=st.from_type(PID),
+        _pid=pid,
         **kwargs,
     )