+++ tests/testing/strategies_test.py
@@ -4,7 +4,7 @@ from dateutil.parser import parse as parse_datetime
 from hypothesis import given, settings
 from hypothesis import strategies as st
 
-from scitacean import DatasetType
+from scitacean import PID, DatasetType
 from scitacean.testing import strategies as sst
 
 
@@ -32,6 +32,8 @@ def test_datasets_can_set_type_to_derived(dset):
         access_groups=["group1", "another-group"],
         source_folder="/the/source/of/truth",
         created_at=parse_datetime("2123-11-29T17:44:56Z"),
+        pid=PID(prefix="prefix", pid="some-id"),
+        description=None,
     )
 )
 def test_datasets_can_fix_fields(dset):
@@ -39,14 +41,19 @@ def test_datasets_can_fix_fields(dset):
     assert dset.access_groups == ["group1", "another-group"]
     assert dset.source_folder == "/the/source/of/truth"
     assert dset.created_at == parse_datetime("2123-11-29T17:44:56Z")
+    assert dset.pid == PID(prefix="prefix", pid="some-id")
+    assert dset.description is None
 
 
 @settings(max_examples=20)
 @given(
     sst.datasets(
-        created_by=st.just("the-creator"), owner=st.sampled_from(("owner1", "owner2"))
+        created_by=st.just("the-creator"),
+        owner=st.sampled_from(("owner1", "owner2")),
+        name=st.text(min_size=2, max_size=5),
     )
 )
 def test_datasets_can_set_strategy_for_fields(dset):
     assert dset.created_by == "the-creator"
     assert dset.owner in ("owner1", "owner2")
+    assert 2 <= len(dset.name) <= 5